+++ langchain-core/src/callbacks/promises.ts
@@ -1,52 +1,3 @@
-import PQueueMod from "p-queue";
-import { AsyncLocalStorageProviderSingleton } from "../singletons/index.js";
+import { awaitAllCallbacks, consumeCallback } from "../singletons/callbacks.js";
 
-let queue: typeof import("p-queue")["default"]["prototype"];
-
-/**
- * Creates a queue using the p-queue library. The queue is configured to
- * auto-start and has a concurrency of 1, meaning it will process tasks
- * one at a time.
- */
-function createQueue() {
-  const PQueue = "default" in PQueueMod ? PQueueMod.default : PQueueMod;
-  return new PQueue({
-    autoStart: true,
-    concurrency: 1,
-  });
-}
-
-/**
- * Consume a promise, either adding it to the queue or waiting for it to resolve
- * @param promiseFn Promise to consume
- * @param wait Whether to wait for the promise to resolve or resolve immediately
- */
-export async function consumeCallback<T>(
-  promiseFn: () => Promise<T> | T | void,
-  wait: boolean
-): Promise<void> {
-  if (wait === true) {
-    // Clear config since callbacks are not part of the root run
-    await AsyncLocalStorageProviderSingleton.runWithConfig(
-      undefined,
-      async () => promiseFn()
-    );
-  } else {
-    if (typeof queue === "undefined") {
-      queue = createQueue();
-    }
-    void queue.add(() =>
-      AsyncLocalStorageProviderSingleton.runWithConfig(undefined, async () =>
-        promiseFn()
-      )
-    );
-  }
-}
-
-/**
- * Waits for all promises in the queue to resolve. If the queue is
- * undefined, it immediately resolves a promise.
- */
-export function awaitAllCallbacks(): Promise<void> {
-  return typeof queue !== "undefined" ? queue.onIdle() : Promise.resolve();
-}
+export { awaitAllCallbacks, consumeCallback };
